Crate oatie

Defines types for using operational tranform to modify HTML-like documents.

See the book for more details: http://tcr.github.io/edit-text/

Modules

apply
Methods to apply an operation to a document.

compose
Composes two operations together.

deserialize
Update legacy serialization formats to a modern format.

doc
Defines document types, operation types, and cursor types.

macros
Macros for more easily creating operations and documents in code.

normalize

place

rtf
Contains types that define the different groups and styles the document tree can contain. This code should live outside of oatie.

schema

stepper
Enables stepping through a span operation.

transform
Performs operational transform on a Doc.

transform_test
Helper methods for performing transform tests.

validate
Validates a document matches a given Schema.

wasm

writer
Classes for generating operation steps (add / del).

Macros

add_span

del_span

doc_span

op

Traits

OT
A type that can have operational transform applied to it. The OT trait is implemented on an operation object, and its associated type Doc is what the operation should operate on.

Schema

Track